[internal] cleanup after LitleNew gateway
