+++ lib/active_merchant/billing/gateways/litle.rb
@@ -3,7 +3,7 @@ require 'nokogiri'
 module ActiveMerchant #:nodoc:
   module Billing #:nodoc:
     class LitleGateway < Gateway
-      SCHEMA_VERSION = '9.4'
+      SCHEMA_VERSION = '9.14'
 
       class_attribute :pre_live_url
 
@@ -13,46 +13,50 @@ module ActiveMerchant #:nodoc:
 
       self.supported_countries = ['US']
       self.default_currency = 'USD'
-      self.supported_cardtypes = [:visa, :master, :american_express, :discover, :diners_club, :jcb]
+      self.supported_cardtypes = %i[visa master american_express discover diners_club jcb]
 
       self.homepage_url = 'http://www.vantiv.com/'
       self.display_name = 'Vantiv eCommerce'
 
-      # Public: Create a new Litle gateway.
-      #
-      # options - A hash of options:
-      #           :login         - The user.
-      #           :password      - The password.
-      #           :merchant_id   - The merchant id.
-      def initialize(options={})
+      def initialize(options = {})
         requires!(options, :login, :password, :merchant_id)
         super
       end
 
-      def purchase(money, payment_method, options={})
+      def purchase(money, payment_method, options = {})
         request = build_xml_request do |doc|
           add_authentication(doc)
-          doc.sale(transaction_attributes(options)) do
-            add_auth_purchase_params(doc, money, payment_method, options)
+          if check?(payment_method)
+            doc.echeckSale(transaction_attributes(options)) do
+              add_echeck_purchase_params(doc, money, payment_method, options)
+            end
+          else
+            doc.sale(transaction_attributes(options)) do
+              add_auth_purchase_params(doc, money, payment_method, options)
+            end
           end
         end
-
-        commit(:sale, request)
+        check?(payment_method) ? commit(:echeckSales, request, money) : commit(:sale, request, money)
       end
 
-      def authorize(money, payment_method, options={})
+      def authorize(money, payment_method, options = {})
         request = build_xml_request do |doc|
           add_authentication(doc)
-          doc.authorization(transaction_attributes(options)) do
-            add_auth_purchase_params(doc, money, payment_method, options)
+          if check?(payment_method)
+            doc.echeckVerification(transaction_attributes(options)) do
+              add_echeck_purchase_params(doc, money, payment_method, options)
+            end
+          else
+            doc.authorization(transaction_attributes(options)) do
+              add_auth_purchase_params(doc, money, payment_method, options)
+            end
           end
         end
-
-        commit(:authorization, request)
+        check?(payment_method) ? commit(:echeckVerification, request, money) : commit(:authorization, request, money)
       end
 
-      def capture(money, authorization, options={})
-        transaction_id, _ = split_authorization(authorization)
+      def capture(money, authorization, options = {})
+        transaction_id, = split_authorization(authorization)
 
         request = build_xml_request do |doc|
           add_authentication(doc)
@@ -63,7 +67,7 @@ module ActiveMerchant #:nodoc:
           end
         end
 
-        commit(:capture, request)
+        commit(:capture, request, money)
       end
 
       def credit(money, authorization, options = {})
@@ -71,19 +75,24 @@ module ActiveMerchant #:nodoc:
         refund(money, authorization, options)
       end
 
-      def refund(money, authorization, options={})
-        transaction_id, _ = split_authorization(authorization)
-
+      def refund(money, payment, options = {})
         request = build_xml_request do |doc|
           add_authentication(doc)
           add_descriptor(doc, options)
-          doc.credit(transaction_attributes(options)) do
-            doc.litleTxnId(transaction_id)
-            doc.amount(money) if money
+          doc.send(refund_type(payment), transaction_attributes(options)) do
+            if payment.is_a?(String)
+              transaction_id, = split_authorization(payment)
+              doc.litleTxnId(transaction_id)
+              doc.amount(money) if money
+            elsif check?(payment)
+              add_echeck_purchase_params(doc, money, payment, options)
+            else
+              add_credit_params(doc, money, payment, options)
+            end
           end
         end
 
-        commit(:credit, request)
+        commit(refund_type(payment), request)
       end
 
       def verify(creditcard, options = {})
@@ -93,13 +102,14 @@ module ActiveMerchant #:nodoc:
         end
       end
 
-      def void(authorization, options={})
-        transaction_id, kind = split_authorization(authorization)
+      def void(authorization, options = {})
+        transaction_id, kind, money = split_authorization(authorization)
 
         request = build_xml_request do |doc|
           add_authentication(doc)
           doc.send(void_type(kind), transaction_attributes(options)) do
             doc.litleTxnId(transaction_id)
+            doc.amount(money) if void_type(kind) == :authReversal
           end
         end
 
@@ -113,6 +123,11 @@ module ActiveMerchant #:nodoc:
             doc.orderId(truncate(options[:order_id], 24))
             if payment_method.is_a?(String)
               doc.paypageRegistrationId(payment_method)
+            elsif check?(payment_method)
+              doc.echeckForToken do
+                doc.accNum(payment_method.account_number)
+                doc.routingNum(payment_method.routing_number)
+              end
             else
               doc.accountNumber(payment_method.number)
               doc.cardValidationNum(payment_method.verification_value) if payment_method.verification_value
@@ -152,25 +167,46 @@ module ActiveMerchant #:nodoc:
       }
 
       AVS_RESPONSE_CODE = {
-          '00' => 'Y',
-          '01' => 'X',
-          '02' => 'D',
-          '10' => 'Z',
-          '11' => 'W',
-          '12' => 'A',
-          '13' => 'A',
-          '14' => 'P',
-          '20' => 'N',
-          '30' => 'S',
-          '31' => 'R',
-          '32' => 'U',
-          '33' => 'R',
-          '34' => 'I',
-          '40' => 'E'
+        '00' => 'Y',
+        '01' => 'X',
+        '02' => 'D',
+        '10' => 'Z',
+        '11' => 'W',
+        '12' => 'A',
+        '13' => 'A',
+        '14' => 'P',
+        '20' => 'N',
+        '30' => 'S',
+        '31' => 'R',
+        '32' => 'U',
+        '33' => 'R',
+        '34' => 'I',
+        '40' => 'E'
       }
 
       def void_type(kind)
-        (kind == 'authorization') ? :authReversal : :void
+        if kind == 'authorization'
+          :authReversal
+        elsif kind == 'echeckSales'
+          :echeckVoid
+        else
+          :void
+        end
+      end
+
+      def refund_type(payment)
+        _, kind, = split_authorization(payment)
+        if check?(payment) || kind == 'echeckSales'
+          :echeckCredit
+        else
+          :credit
+        end
+      end
+
+      def check?(payment_method)
+        return false if payment_method.is_a?(String)
+
+        card_brand(payment_method) == 'check'
       end
 
       def add_authentication(doc)
@@ -189,7 +225,39 @@ module ActiveMerchant #:nodoc:
         add_payment_method(doc, payment_method, options)
         add_pos(doc, payment_method)
         add_descriptor(doc, options)
+        add_merchant_data(doc, options)
         add_debt_repayment(doc, options)
+        add_stored_credential_params(doc, options)
+      end
+
+      def add_credit_params(doc, money, payment_method, options)
+        doc.orderId(truncate(options[:order_id], 24))
+        doc.amount(money)
+        add_order_source(doc, payment_method, options)
+        add_billing_address(doc, payment_method, options)
+        add_payment_method(doc, payment_method, options)
+        add_pos(doc, payment_method)
+        add_descriptor(doc, options)
+        add_merchant_data(doc, options)
+      end
+
+      def add_merchant_data(doc, options = {})
+        if options[:affiliate] || options[:campaign] || options[:merchant_grouping_id]
+          doc.merchantData do
+            doc.affiliate(options[:affiliate]) if options[:affiliate]
+            doc.campaign(options[:campaign]) if options[:campaign]
+            doc.merchantGroupingId(options[:merchant_grouping_id]) if options[:merchant_grouping_id]
+          end
+        end
+      end
+
+      def add_echeck_purchase_params(doc, money, payment_method, options)
+        doc.orderId(truncate(options[:order_id], 24))
+        doc.amount(money)
+        add_order_source(doc, payment_method, options)
+        add_billing_address(doc, payment_method, options)
+        add_payment_method(doc, payment_method, options)
+        add_descriptor(doc, options)
       end
 
       def add_descriptor(doc, options)
@@ -210,10 +278,18 @@ module ActiveMerchant #:nodoc:
           doc.token do
             doc.litleToken(payment_method)
             doc.expDate(options[:expiration_date]) if options[:expiration_date].present?
+            doc.expDate(format_exp_date(options[:basis_expiration_month], options[:basis_expiration_year])) if options[:basis_expiration_month] && options[:basis_expiration_year]
           end
         elsif payment_method.respond_to?(:track_data) && payment_method.track_data.present?
           doc.card do
             doc.track(payment_method.track_data)
+          end
+        elsif check?(payment_method)
+          doc.echeck do
+            doc.accType(payment_method.account_type.capitalize)
+            doc.accNum(payment_method.account_number)
+            doc.routingNum(payment_method.routing_number)
+            doc.checkNum(payment_method.number) if payment_method.number
           end
         else
           doc.card do
@@ -226,23 +302,68 @@ module ActiveMerchant #:nodoc:
             doc.cardholderAuthentication do
               doc.authenticationValue(payment_method.payment_cryptogram)
             end
+          elsif options[:order_source]&.start_with?('3ds')
+            doc.cardholderAuthentication do
+              doc.authenticationValue(options[:cavv]) if options[:cavv]
+              doc.authenticationTransactionId(options[:xid]) if options[:xid]
+            end
           end
         end
       end
 
-      def add_billing_address(doc, payment_method, options)
-        doc.billToAddress do
-          if payment_method.respond_to?(:name)
-            doc.name(payment_method.name)
-          elsif options[:billing_address].has_key?(:first_name) || options[:billing_address].has_key?(:last_name)
-            doc.name("#{options[:billing_address][:first_name]} #{options[:billing_address][:last_name]}".strip)
-          end
+      def add_stored_credential_params(doc, options = {})
+        return unless options[:stored_credential]
 
-          if options[:email]
-            doc.email(options[:email])
-          elsif options[:billing_address] && options[:billing_address][:email]
-            doc.email(options[:billing_address][:email])
+        if options[:stored_credential][:initial_transaction]
+          add_stored_credential_params_initial(doc, options)
+        else
+          add_stored_credential_params_used(doc, options)
+        end
+      end
+
+      def add_stored_credential_params_initial(doc, options)
+        case options[:stored_credential][:reason_type]
+        when 'unscheduled'
+          doc.processingType('initialCOF')
+        when 'installment'
+          doc.processingType('initialInstallment')
+        when 'recurring'
+          doc.processingType('initialRecurring')
+        end
+      end
+
+      def add_stored_credential_params_used(doc, options)
+        if options[:stored_credential][:reason_type] == 'unscheduled'
+          if options[:stored_credential][:initiator] == 'merchant'
+            doc.processingType('merchantInitiatedCOF')
+          else
+            doc.processingType('cardholderInitiatedCOF')
           end
+        end
+        doc.originalNetworkTransactionId(options[:stored_credential][:network_transaction_id])
+      end
+
+      def skip_billing_address?(payment_method, options)
+        return false if options[:email]
+        return true if payment_method.is_a?(String)
+
+        false
+      end
+
+      def add_billing_address(doc, payment_method, options)
+        return if skip_billing_address?(payment_method, options)
+
+        doc.billToAddress do
+          if check?(payment_method)
+            doc.name(payment_method.name)
+            doc.firstName(payment_method.first_name)
+            doc.lastName(payment_method.last_name)
+          elsif !payment_method.is_a?(String)
+            doc.name(payment_method.name)
+          else
+            doc.name(options[:full_name])
+          end
+          doc.email(options[:email]) if options[:email]
 
           add_address(doc, options[:billing_address])
         end
@@ -270,15 +391,47 @@ module ActiveMerchant #:nodoc:
       end
 
       def add_order_source(doc, payment_method, options)
-        if options[:order_source]
-          doc.orderSource(options[:order_source])
+        order_source = order_source(options)
+        if order_source
+          doc.orderSource(order_source)
         elsif payment_method.is_a?(NetworkTokenizationCreditCard) && payment_method.source == :apple_pay
           doc.orderSource('applepay')
+        elsif payment_method.is_a?(NetworkTokenizationCreditCard) && payment_method.source == :android_pay
+          doc.orderSource('androidpay')
         elsif payment_method.respond_to?(:track_data) && payment_method.track_data.present?
           doc.orderSource('retail')
         else
           doc.orderSource('ecommerce')
         end
+      end
+
+      def order_source(options = {})
+        return options[:order_source] unless options[:stored_credential]
+
+        order_source = nil
+
+        case options[:stored_credential][:reason_type]
+        when 'unscheduled'
+          if options[:stored_credential][:initiator] == 'merchant'
+            # For merchant-initiated, we should always set order source to
+            # 'ecommerce'
+            order_source = 'ecommerce'
+          else
+            # For cardholder-initiated, we rely on #add_order_source's
+            # default logic to set orderSource appropriately
+            order_source = options[:order_source]
+          end
+        when 'installment'
+          order_source = 'installment'
+        when 'recurring'
+          order_source = 'recurring'
+        end
+
+        if order_source == 'recurring' && options[:stored_credential][:initial_transaction]
+          order_source = 'ecommerce'
+        end
+
+        order_source
       end
 
       def add_pos(doc, payment_method)
@@ -292,7 +445,11 @@ module ActiveMerchant #:nodoc:
       end
 
       def exp_date(payment_method)
-        "#{format(payment_method.month, :two_digits)}#{format(payment_method.year, :two_digits)}"
+        format_exp_date(payment_method.month, payment_method.year)
+      end
+
+      def format_exp_date(month, year)
+        "#{format(month, :two_digits)}#{format(year, :two_digits)}"
       end
 
       def parse(kind, xml)
@@ -300,7 +457,7 @@ module ActiveMerchant #:nodoc:
 
         doc = Nokogiri::XML(xml).remove_namespaces!
         doc.xpath("//litleOnlineResponse/#{kind}Response/*").each do |node|
-          if (node.elements.empty?)
+          if node.elements.empty?
             parsed[node.name.to_sym] = node.text
           else
             node.elements.each do |childnode|
@@ -312,21 +469,21 @@ module ActiveMerchant #:nodoc:
 
         if parsed.empty?
           %w(response message).each do |attribute|
-            parsed[attribute.to_sym] = doc.xpath("//litleOnlineResponse").attribute(attribute).value
+            parsed[attribute.to_sym] = doc.xpath('//litleOnlineResponse').attribute(attribute).value
           end
         end
 
         parsed
       end
 
-      def commit(kind, request)
+      def commit(kind, request, money = nil)
         parsed = parse(kind, ssl_post(url, request, headers))
 
         options = {
-          authorization: authorization_from(kind, parsed),
+          authorization: authorization_from(kind, parsed, money),
           test: test?,
-          :avs_result => { :code => AVS_RESPONSE_CODE[parsed[:fraudResult_avsResult]] },
-          :cvv_result => parsed[:fraudResult_cardValidationResult]
+          avs_result: { code: AVS_RESPONSE_CODE[parsed[:fraudResult_avsResult]] },
+          cvv_result: parsed[:fraudResult_cardValidationResult]
         }
 
         Response.new(success_from(kind, parsed), parsed[:message], parsed, options)
@@ -334,16 +491,17 @@ module ActiveMerchant #:nodoc:
 
       def success_from(kind, parsed)
         return (parsed[:response] == '000') unless kind == :registerToken
+
         %w(000 801 802).include?(parsed[:response])
       end
 
-      def authorization_from(kind, parsed)
-        (kind == :registerToken) ? parsed[:litleToken] : "#{parsed[:litleTxnId]};#{kind}"
+      def authorization_from(kind, parsed, money)
+        kind == :registerToken ? parsed[:litleToken] : "#{parsed[:litleTxnId]};#{kind};#{money}"
       end
 
       def split_authorization(authorization)
-        transaction_id, kind = authorization.to_s.split(';')
-        [transaction_id, kind]
+        transaction_id, kind, money = authorization.to_s.split(';')
+        [transaction_id, kind, money]
       end
 
       def transaction_attributes(options)
@@ -351,7 +509,7 @@ module ActiveMerchant #:nodoc:
         attributes[:id] = truncate(options[:id] || options[:order_id], 24)
         attributes[:reportGroup] = options[:merchant] || 'Default Report Group'
         attributes[:customerId] = options[:customer]
-        attributes.delete_if { |key, value| value == nil }
+        attributes.delete_if { |_key, value| value == nil }
         attributes
       end
 
@@ -359,7 +517,7 @@ module ActiveMerchant #:nodoc:
         {
           merchantId: @options[:merchant_id],
           version: SCHEMA_VERSION,
-          xmlns: "http://www.litle.com/schema"
+          xmlns: 'http://www.litle.com/schema'
         }
       end
 
@@ -371,10 +529,6 @@ module ActiveMerchant #:nodoc:
         builder.doc.root.to_xml
       end
 
-      def has_credentials?
-        @options[:login].present? && @options[:password].present?
-      end
-
       def url
         if test? && has_credentials?
           pre_live_url
@@ -383,6 +537,10 @@ module ActiveMerchant #:nodoc:
         else
           live_url
         end
+      end
+
+      def has_credentials?
+        @options[:login].present? && @options[:password].present?
       end
 
       def headers
